Fix crash when inserting nil as KVC value
When mapping the values, Motis' validator returns that nil is valid (and it might be the case somewhere) but when you are calling setValue:forKey: value can't be nil, so in this case we have to check for that too. If it gets through as nil it's simply going to crash the app.